dev: Populate the Pipfile for development
The [docs] recommend that libraries use `setup.py`'s `install_requires` field for managing real
library dependencies and the Pipfile for development dependencies. We already set the required
version of `requests` in `setup.py`, so this sets up the Pipfile with only development
dependencies.

To avoid too much change from `requirements.txt`, this includes `stytch` as an editable dependency
in the Pipfile. So `pipenv install` should be a usable replacement for
`pip install -r requirements.txt`.

[docs]: https://pipenv.pypa.io/en/latest/advanced/#pipfile-vs-setup-py