fix: Correct mypy errors (#62)
We'd like to be able to rely on mypy to validate our type annotations, but this repo currently has
some type-checking issues. This adds a mypy step to the CI pipeline and fixes all of those type
issues so we can start enforcing it on PRs.

I chose to type all request bodies (`data` variables) as `Dict[str, Any]` because that was the
easiest way to get `{"attributes": {...}}` to fit without an unwieldy `Union`. We can narrow these
types in the future if we want better constraints on individual requests.

While I was here, I happened across a [mutable default value], so I changed that to avoid weird
behavior.

[mutable default value]: https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments