feat: Allow custom env values for internal dev URLs (#73)
API client setup uses the `env` value to set the correct base URI for that environment. This is
great for the production live and test environments (which are the only supported ones!), but that
makes it difficult to use this within a Stytch development environment as a Stytch API developer.

So this relaxes the error on the `env` value. If it's one of the two recognized environments,
derive the URL from it. If not, treat it as a base URL to use when building requests.